+++ src/baas/configs/application.yaml
@@ -1,13 +1,6 @@
 # 基础配置
 app_name: "secbaas"
-enable_sidecar: true
 workers: 8
-
-# Sidecar配置
-sidecar_config:
-  host: "localhost"
-  max_send_message_length: 4194304 # 4MB
-  max_receive_message_length: 4194304 # 4MB
 
 log_config:
   trace_log_dir: ""
@@ -18,9 +11,6 @@ log_config:
 module_config:
   web:
     port: 8888 # 修改端口需要同步修改nginx.conf内配置
-    workers: 8
-    threads: 8
-    start: "secbaas.community.adapters.web.app:app"
 
 user_config:
   # 插件配置
@@ -42,19 +32,6 @@ user_config:
       k8s: stub # real or stub
       docker: stub # real or stub
       poolab: stub # real or stub
-  # K8s 模板配置示例（用于 K8s PaaS 平台设备创建）
-  # 实际配置从数据库 baas_device_template.config 字段读取，此处仅为环境级参考。
-  k8s_template_config:
-    # kubeconfig: |
-    #   apiVersion: v1
-    #   kind: Config
-    #   clusters: [...]
-    namespace: "default"
-    image: "registry.example.com/bot-runtime:latest"
-    cpu_request: "500m"
-    cpu_limit: "1"
-    memory_request: "512Mi"
-    memory_limit: "1Gi"
   # Arca 沙箱配置（用于定时任务续期）
   arca:
     enabled: true  # 设置为 true 启用沙箱续期定时任务
