--- src/baas/sqls/ac_bot_publish.sql
@@ -1,25 +1,0 @@
-CREATE TABLE IF NOT EXISTS `ac_bot_publish` (
-  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
-  `source_bot_pk` bigint(20) unsigned NOT NULL COMMENT 'ac_bots 表主键',
-  `source_bot_id` varchar(512) NOT NULL COMMENT '主要用来区分： 数字分身/个人',
-  `publish_bot_id` varchar(1024) NOT NULL COMMENT '发布拼的bot_id=source_bot_id + pub + version',
-  `name` varchar(2048) NOT NULL COMMENT 'bot名',
-  `description` varchar(4096) DEFAULT NULL COMMENT '描述',
-  `owner_id` varchar(128) NOT NULL COMMENT '工号',
-  `owner_name` varchar(128) DEFAULT NULL COMMENT '花名',
-  `status` varchar(128) NOT NULL COMMENT '发布状态：init，sucess， release，fail',
-  `version` bigint(20) unsigned DEFAULT NULL COMMENT '版本号',
-  `last_pub_id` bigint(20) unsigned NOT NULL COMMENT '上次发布成功id',
-  `env` varchar(64) NOT NULL COMMENT 'prod/pre/dev',
-  `ext` varchar(60000) DEFAULT NULL COMMENT '扩展信息',
-  `gmt_create` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
-  `gmt_modified` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '修改时间',
-  `permission_owner` varchar(64) NOT NULL COMMENT '权限归属',
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `uk_oi_p_b_v` (`owner_id`, `publish_bot_id`, `version`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_l_p_i` (`last_pub_id`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_owner_name` (`owner_name`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_name` (`name`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_pb_owner_s` (`publish_bot_id`, `owner_id`, `status`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_sbp_e_s` (`source_bot_pk`, `env`, `status`) BLOCK_SIZE 16384 GLOBAL
-);
