--- src/baas/sqls/ac_bots.sql
@@ -1,34 +1,0 @@
-CREATE TABLE IF NOT EXISTS `ac_bots` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT COMMENT 'id',
-  `bot_id` varchar(256) NOT NULL COMMENT 'Bot ID: yyyymmdd + random 8 digits',
-  `bot_name` varchar(1024) DEFAULT NULL COMMENT 'Bot name',
-  `bot_desc` varchar(1024) DEFAULT NULL COMMENT 'Bot description',
-  `entity_id` varchar(1024) NOT NULL COMMENT 'Entity ID (e.g., staff_xxx, proj_xxx, team_xxx)',
-  `entity_type` varchar(1024) NOT NULL COMMENT 'Entity type: staff, proj, team',
-  `creator_id` varchar(1024) NOT NULL COMMENT 'Creator user ID',
-  `owner_id` varchar(1024) NOT NULL COMMENT 'Owner user ID (set to creator_id on creation)',
-  `engine_types` varchar(2048) DEFAULT NULL COMMENT 'Engine types (JSON string list)',
-  `status` varchar(50) NOT NULL DEFAULT 'PENDING' COMMENT 'Bot status: PENDING, ACTIVE, FAILED, RELEASED',
-  `binding_id` bigint(20) DEFAULT NULL COMMENT 'Associated device binding ID',
-  `gmt_create` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Creation time',
-  `gmt_modified` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'Last modification time',
-  `modifier_id` varchar(1024) DEFAULT NULL COMMENT 'Last modifier user ID',
-  `share_policy` varchar(2048) DEFAULT NULL COMMENT 'Sharing configuration (JSON string)',
-  `is_delete` tinyint(1) NOT NULL DEFAULT '0' COMMENT 'Soft delete flag: 0=active, 1=deleted',
-  `active_engine` varchar(256) DEFAULT NULL COMMENT 'active_engine',
-  `device_id` varchar(256) DEFAULT NULL COMMENT 'device_id',
-  `env` varchar(20) DEFAULT NULL COMMENT '环境标记',
-  `owner_name` varchar(100) DEFAULT NULL COMMENT 'onwer花名',
-  `public` varchar(64) NOT NULL DEFAULT '0' COMMENT '是否公开',
-  `ext` text DEFAULT NULL COMMENT '扩展字段',
-  `bot_type` varchar(128) DEFAULT 'personal' COMMENT 'personal/service',
-  PRIMARY KEY (`id`),
-  KEY `idx_entity` (`entity_id`, `entity_type`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_creator` (`creator_id`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_owner` (`owner_id`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_status` (`status`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_binding_id` (`binding_id`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_is_delete` (`is_delete`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_device_id` (`device_id`) BLOCK_SIZE 16384 GLOBAL,
-  KEY `idx_p_bn_on` (`public`, `bot_name`, `owner_name`) BLOCK_SIZE 16384 GLOBAL
-) COMMENT = 'Bot information table';
