+++ src/baas/src/secbaas/community/core/service/sse/__init__.py
@@ -3,12 +3,10 @@
 提供 StreamConverter 的工厂和默认实现。
 """
 
-from ._bcn_converter import BcnStreamConverter
 from ._default_converter import DefaultStreamConverter
 from ._registry import SseConverterFactory
 
 __all__ = [
-    "BcnStreamConverter",
     "DefaultStreamConverter",
     "SseConverterFactory",
 ]
